refactor: separate tests from src to avoid structure coupling
