+++ packages/skott/test/ecmascript/graph.spec.ts
@@ -1,0 +1,256 @@
+import { expect } from "chai";
+import * as memfs from "memfs";
+
+import { FileReader } from "../../src/filesystem/file-reader.js";
+import { Skott, SkottStructure } from "../../src/skott";
+
+import { mountFakeFileSystem } from "./shared";
+import { buildSkottProjectUsingInMemoryFileExplorer } from "./shared.js";
+
+class InMemoryFileReaderWithFakeStats implements FileReader {
+  read(filename: string): Promise<string> {
+    return new Promise((resolve) => {
+      /* eslint-disable no-sync */
+      resolve(memfs.fs.readFileSync(filename, "utf-8") as string);
+    });
+  }
+  stats(filename: string): Promise<number> {
+    return new Promise((resolve) => {
+      memfs.fs.stat(filename, (err, data) => {
+        if (err) {
+          return resolve(0);
+        }
+
+        if (data) {
+          return resolve(Number(data.size));
+        }
+
+        return resolve(0);
+      });
+    });
+  }
+}
+
+async function makeScott(): Promise<SkottStructure> {
+  const skott = new Skott(
+    {
+      entrypoint: "index.js",
+      circularMaxDepth: Number.POSITIVE_INFINITY,
+      includeBaseDir: false,
+      dependencyTracking: {
+        thirdParty: true,
+        builtin: true
+      }
+    },
+    new InMemoryFileReaderWithFakeStats()
+  );
+  const skottInstance = await skott.initialize();
+
+  return skottInstance.getStructure();
+}
+
+describe("When building the project structure independently of JavaScript or TypeScript", () => {
+  describe("When searching for leaves", () => {
+    it("should only return nodes identifiers with no children", async () => {
+      mountFakeFileSystem({
+        "index.js": "export * from './file1.js';",
+        "file1.js": "import { F2 } from './file2.js';",
+        "file2.js": "export const F2 = 10;"
+      });
+
+      const { leaves } = await buildSkottProjectUsingInMemoryFileExplorer(
+        "index.js"
+      );
+
+      expect(leaves).to.be.deep.equal(["file2.js"]);
+    });
+  });
+
+  describe("When deeply searching for parents of a given node", () => {
+    it("should traverse the graph bottom-to-top and collect every file depending on the target", async () => {
+      mountFakeFileSystem({
+        "a.js": `
+              import { b } from "./b.js";
+              export const a = () => b();
+            `,
+        "b.js": `
+              import { c } from "./c.js";
+              export const b = { doSomething: () => bar() };
+            `,
+        "c.js": `
+              import { d } from "./d.js";
+              import { x } from "./x.js";
+              export const c = { doSomething: () => {} };
+            `,
+        "x.js": `
+              export const x = { doSomething: () => {} };
+            `,
+        "d.js": `
+              export const d = { doSomething: () => {} };
+            `
+      });
+
+      const skott = new Skott(
+        {
+          entrypoint: "a.js",
+          circularMaxDepth: Number.POSITIVE_INFINITY,
+          includeBaseDir: false,
+          dependencyTracking: {
+            thirdParty: false,
+            builtin: false
+          }
+        },
+        new InMemoryFileReaderWithFakeStats()
+      );
+
+      const skottInstance = await skott.initialize();
+      expect(skottInstance.findParentsOf("d.js")).to.deep.equal([
+        "c.js",
+        "b.js",
+        "a.js"
+      ]);
+      expect(skottInstance.findParentsOf("c.js")).to.deep.equal([
+        "b.js",
+        "a.js"
+      ]);
+      expect(skottInstance.findParentsOf("a.js")).to.deep.equal([]);
+    });
+  });
+
+  describe("Skott Node details", () => {
+    describe("When there is only one file in the project", () => {
+      describe("When the file is empty", () => {
+        it("should generate basic data", async () => {
+          mountFakeFileSystem({
+            "index.js": ""
+          });
+
+          const { graph } = await makeScott();
+
+          expect(graph["index.js"]).to.deep.equal({
+            id: "index.js",
+            adjacentTo: [],
+            body: {
+              size: 0,
+              thirdPartyDependencies: [],
+              builtinDependencies: []
+            }
+          });
+        });
+      });
+
+      describe("When the file contains some source code", () => {
+        describe("When computing the file size", () => {
+          it("should return the size in bytes of the file", async () => {
+            const fakeFs = {
+              "index.js": `
+                import * as lib from "./lib.js";
+                const x = 3;
+            `,
+              "lib.js": `
+                function doSomething() {
+                    return "doSomething";
+                }
+
+                /**
+                 * next-gen js-doc
+                 */
+                function helloEverybody() {
+                    return "helloEverybody";
+                }
+
+                export { doSomething, helloEverybody };
+            `
+            };
+
+            mountFakeFileSystem(fakeFs);
+
+            const { graph } = await makeScott();
+
+            const indexFile = graph["index.js"];
+            const libFile = graph["lib.js"];
+
+            expect(indexFile.body).to.deep.equal({
+              size: Buffer.byteLength(fakeFs["index.js"], "utf-8"),
+              thirdPartyDependencies: [],
+              builtinDependencies: []
+            });
+            expect(libFile.body).to.deep.equal({
+              size: Buffer.byteLength(fakeFs["lib.js"], "utf-8"),
+              thirdPartyDependencies: [],
+              builtinDependencies: []
+            });
+          });
+        });
+
+        describe("When tracking third-party dependencies", async () => {
+          it("should find one third-party dependency", async () => {
+            mountFakeFileSystem({
+              "index.js": `
+              import * as anything from "./lib.js";
+              import { parseScript } from 'meriyah';
+              import path from "path";
+              import * as fs from "node:fs"; 
+            `,
+              "lib.js": ""
+            });
+
+            const { graph } = await makeScott();
+            const indexFile = graph["index.js"];
+
+            expect(indexFile.body.thirdPartyDependencies).to.deep.equal([
+              "meriyah"
+            ]);
+          });
+          it("should find many third-party dependencies", async () => {
+            memfs.vol.fromJSON(
+              {
+                "index.js": `
+                import * as anything from "./lib.js";
+                import { parseScript } from 'meriyah';
+                import 'side-effect-library';
+                import { getStrategy } from "@nodesecure/vulnera";
+              `,
+                "lib.js": ""
+              },
+              "./"
+            );
+
+            const { graph } = await makeScott();
+            const indexFile = graph["index.js"];
+
+            expect(indexFile.body.thirdPartyDependencies).to.deep.equal([
+              "meriyah",
+              "side-effect-library",
+              "@nodesecure/vulnera"
+            ]);
+          });
+        });
+
+        describe("When tracking native dependencies", async () => {
+          describe("When the source file contains native dependencies", async () => {
+            it("should collect all native dependencies", async () => {
+              mountFakeFileSystem({
+                "index.js": `
+                import * as anything from "./lib.js";
+                import { parseScript } from 'meriyah';
+                import path from "path";
+                import * as fs from "node:fs"; 
+              `,
+                "lib.js": ""
+              });
+
+              const { graph } = await makeScott();
+              const indexFile = graph["index.js"];
+
+              expect(indexFile.body.builtinDependencies).to.deep.equal([
+                "path",
+                "node:fs"
+              ]);
+            });
+          });
+        });
+      });
+    });
+  });
+});
