+++ packages/skott/tsconfig.json
@@ -1,4 +1,4 @@
 {
   "extends": "@skott/config/tsconfig.json",
-  "include": ["src/**/*.ts", "bin/**/*.ts", "index.ts"]
+  "include": ["src/**/*.ts", "bin/**/*.ts", "test/**/*.ts", "index.ts"]
 }
